src/requirements.txt
# AWS Integration
boto3
botocore

# Database
psycopg2-binary
pgvector
sqlalchemy

# Data Processing
numpy
pandas
openpyxl

# Configuration and Utilities
configparser
colorama
tqdm
tiktoken

# JSON Processing
json5

# File System Operations
fsspec
PyPDF2
python-docx
python-pptx
docx
openpyxl

# Date and Time Handling
python-dateutil
psutil

#Langchain
langchain-core
langchain
langchain-postgres
langchain_community
requests

#Environment
python-dotenv
pythonnet
dotnetcore2
langchain
h11
pip
attrs
distro
pycparser
cffi
pytz
PyYAML
idna
multidict
propcache
sniffio
s3transfer
jmespath
httpcore
httpx
zstandard


charset-normalizer
typing_extensions
aiohttp
yarl
orjson
MarkupSafe
aiosignal
frozenlist
aiohappyeyeballs
asyncpg
certifi
psycopg
psycopg-pool
urllib3
six
greenlet
anyio
et_xmlfile
pydantic
pydantic_core
annotated-types
tenacity
langsmith
langchain-text-splitters
requests-toolbelt
packaging
clr_loader
marshmallow
jsonpointer
jsonpatch
typing-inspect
dataclasses-json
pydantic-settings
mypy_extensions
httpx-sse
tiktoken

#API
fastapi